changing behavior when fetching generic paths
Previous implementation defined the generic path at constructor time.
During testing, that required intervening in the environment variables
before the class-under-test was loaded by the JVM, which worked fine
when the test ran in isolation but broke when the test was run as part
of a suite (i.e., when running all tests in `src/test/java`)

Using Suppliers to resolve those paths when the get() method is called
has no real purpose during program execution, and only affects the WIN
environment for now, but it does solve the above problem during
cross-platform testing, and may prove to have other uses in future. I
can't see any argument AGAINST it really. If somebody wanted to be sure
something was fully resolved at constructor-time, they could still just
provide a Supplier of a static value, which is more or less what the
non-WIN, non-ALLUSERS paths are doing.