Adding base ingester class plus test harness
Base class for any class that manages ingesting LauncherOptions from an
external source. Concrete implementations pending are for reading in
command-line arguments, reading in an embedded Java properties file, and
reading a JSON object from a remote endpoint, but any number of other
implementations are possible (reading from S3? Redis? DynamoDB? why not?)

note no concrete tests yet, as the base class is abstract and doesn't do
anything by itself. But there are test objects, assertions, etc. that
the concrete classes can use to assert that the superclass has done what
it is supposed to. This way that logic doesn't have to be duplicated
across concrete implementations.